[chore]: refactor to make use of custom hooks
